requirements.txt
async-generator==1.10
attrs==23.1.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
cryptography==40.0.2
docopt==0.6.2
exceptiongroup==1.1.1
h11==0.14.0
idna==3.4
outcome==1.2.0
pdfminer.six==20220524
py-pdf-parser==0.10.2
pycparser==2.21
PySocks==1.7.1
requests==2.28.2
selenium==4.8.3
sniffio==1.3.0
sortedcontainers==2.4.0
trio==0.22.0
trio-websocket==0.10.2
undetected-chromedriver==3.4.6
urllib3==1.26.15
Wand==0.6.10
websockets==11.0.2
wsproto==1.2.0
